Intermediate result files now have a structured name (sample number + site lon/lat).
Files are also prepended with an underscore; once the file is closed,
the underscore is removed. (This is used as an indicator to prevent
collector code from trying to read the file while it is still being written to.)
Collectors should only read files if the leading underscore is not present.


Former-commit-id: 070aa41c880ea891143d079c8fe375d2244aa04b [formerly 761a157703409840b22fa08e68ec8ada92f28fda]
Former-commit-id: 793a1d8cf4a136395532198abdbd4e4cb2c7db34