attrel/base: redone CoeffsTable (+)
now dropped SACoeffsTable and all the chaining business. instead tune
everything for special-case tables (which apparently are not
special-cases) that contain sa periods for only one damping value and
mix SA with other IMTs in a way that SA values are specified by period
value while others just by name. this is very much in sync with what can
be found in many (if not all) GMPE papers and it works much faster than
traversing chained tables.