gsim/base: more efficent context creation and poes calculation (+)
now the prepare_distances() function is gone, make_context() doesn't
expect precalculated distances anymore because it's now called
make_contexts() and it takes a list of sites, so it runs distance
calculation itself. in turn get_poes expects a list of contexts and
return a mapping of IMTs to 2d arrays (sites x IMLs). those changes
result in significant reduction of number of function calls and
therefore the code performs faster.