Hack to handle recent changes to nrml submodule
The nrml project is not yet treated like its own package... but it should be.

Since we just added some code to nrml, this causes problems with imports in
the code which sits in the nrml submodule while running oq-engine tests.
The easy workaround is to add openquake/nrml (the location of the submodule)
to the PYTHONPATH in test scripts.


Former-commit-id: 88d9b1b2db4f030de9d2192d0f60f901ff3ed460 [formerly 61234e06f4280b1e1d363cd52a7a1e08d43d3ddd] [formerly 61234e06f4280b1e1d363cd52a7a1e08d43d3ddd [formerly 7786b56c771ed97df562c001f0688d5dc2ace243]]
Former-commit-id: 196730742a9646f457426a02f8826fd10ad98da7
Former-commit-id: fbe8ea5b73705b1a2331e2a2418f4217cf788b10