Hack to handle recent changes to nrml submodule
The nrml project is not yet treated like its own package... but it should be.

Since we just added some code to nrml, this causes problems with imports in
the code which sits in the nrml submodule while running oq-engine tests.
The easy workaround is to add openquake/nrml (the location of the submodule)
to the PYTHONPATH in test scripts.