Hack to handle recent changes to nrml submodule
The nrml project is not yet treated like its own package... but it should be.

Since we just added some code to nrml, this causes problems with imports in
the code which sits in the nrml submodule while running oq-engine tests.
The easy workaround is to add openquake/nrml (the location of the submodule)
to the PYTHONPATH in test scripts.


Former-commit-id: 6495ef0d2036cf56b724cf6875ff5dfc8de70487 [formerly dc939689b3101dd4528fac1c9020dbf6156541a8] [formerly dc939689b3101dd4528fac1c9020dbf6156541a8 [formerly 06390a82c2f66b4bdcc1098b88b14b842492bca2]]
Former-commit-id: 33b983f9d92d69cd58648b544b6fa204e05b068c
Former-commit-id: 59881ab0cd17bb1b13b027ca44245c8f88b70c70