Function for storing a `site model` in the DB
store_site_model in openquake/calculators/hazard/general.py stores
a structured version of a site model into the database, given an XML
source. The parser used to read the XML structures location data such
that the site model records created can be subject to geospatial
queries.

I added a basic test for this function. Next, I need to make it is
used properly in the broader job/calculation context.


Former-commit-id: a4aac75b13c2d92ca83425be7eb6098754b370a8 [formerly a4aac75b13c2d92ca83425be7eb6098754b370a8 [formerly b53e1d068d41517de80853557bf3141e19903a61]]
Former-commit-id: 896e157b12e67be10f7730e8d52a0c320af283ff
Former-commit-id: 344de45dce91ad3261935fc3a239f0ca8b70fc0a