Function for storing a `site model` in the DB
store_site_model in openquake/calculators/hazard/general.py stores
a structured version of a site model into the database, given an XML
source. The parser used to read the XML structures location data such
that the site model records created can be subject to geospatial
queries.

I added a basic test for this function. Next, I need to make it is
used properly in the broader job/calculation context.


Former-commit-id: 8cd5ed52913346092783ca910db29ef96cc9c46c [formerly 8cd5ed52913346092783ca910db29ef96cc9c46c [formerly a5a2f133537b3671e4a8cae2c0347e487481c096]]
Former-commit-id: c3e5e410642cd01551f0c11961c7f0bc2ecddb0f
Former-commit-id: fd38e27d40b7238c598ace8a2cba65700ec3f310