prepare_job_context utility
Added a prepare_job_context util function which takes a path to a config
file and spits back a openquake.engine.JobContext object, which is basically
the complete context required to run a calculation.

This job context can be directly passed to a calculator constructor; this makes
calculator testing set up much easier.


Former-commit-id: f14c6c3267f414483389c116445a259ac15aa03d