prepare_job_context utility
Added a prepare_job_context util function which takes a path to a config
file and spits back a openquake.engine.JobContext object, which is basically
the complete context required to run a calculation.

This job context can be directly passed to a calculator constructor; this makes
calculator testing set up much easier.


Former-commit-id: 1017c0870f269aa542bdadda2501f4ba472b649d [formerly f14c6c3267f414483389c116445a259ac15aa03d]
Former-commit-id: 459bfd2b7d7ceaf21dfbee6e73c3d9ff979515af