Source model dist. probabilities use `decimal`
The nodal plane and hypo depth distributions modelled in the objects
produced by the source model parser now use `decimal.Decimal` types
to represent probablities (instead of `float`). This way, we can
ensure that the probabilities in a distribution sum to 1. This is
necessary because in python `sum([0.1] * 10) != 1.0` (instead, the
sum is 0.9999999999999999).