Fixed a small error in run_tests
Test gathering was missing a handful of tests: some doctests and one
test that we are picking up from the nrml submodule.

I don't like the fact that we're grabbing this nrml test, but once
we start using nrml as a separate library (properly), the submodule can
be removed and this problem will go away.


Former-commit-id: 308c152af424388ffea562cedde303dc8e6837ae [formerly 308c152af424388ffea562cedde303dc8e6837ae [formerly ecede319f93a0346e50246f706a95c1170a006a2]]
Former-commit-id: 8916ab330893e08da4881edb581309943f26aad8
Former-commit-id: 05053b16a3fab531ead963632af97f07b9126e17