Fixed a small error in run_tests
Test gathering was missing a handful of tests: some doctests and one
test that we are picking up from the nrml submodule.

I don't like the fact that we're grabbing this nrml test, but once
we start using nrml as a separate library (properly), the submodule can
be removed and this problem will go away.