Added special handling to haz job profile for geom
With this change, users can specify coordinate pairs for geometry in config
file, which can be parsed with a ConfigParser and passed directly to the model
(which can itself be passed to a model form for validation).

This is nice because we don't have to do a bunch of type conversion; we can go straight
from the config to the model.

We'll be able to delete a TON of parameter and validation code soon.