Sketched out the rest of the execution code
The calc prep looks good now, and we can actually distribute tasks and calculation.

However, there's still a problem on the worker side, dealing with the result
update transaction:
    AttributeError: 'Manager' object has no attribute 'select_for_update'

Since we're actually getting to this point, that means that the calculation
(using the nhlib hazard curve calc) concluded "successfully". This is good
progress.


Former-commit-id: 09b508f7c8233589afcf60ab6d376b23f510af35 [formerly da76ac170d5702c98f23e8741f64a7f53b553c06] [formerly da76ac170d5702c98f23e8741f64a7f53b553c06 [formerly 87e68904fa86a8e531fc6c29d3981b829dfcc1d6]]
Former-commit-id: d9692cf1807a18e36b6ed867a46e3d783a2403a7
Former-commit-id: 87138c87cdb07defe0514edf429b7a3799f3543a