Sketched out the rest of the execution code
The calc prep looks good now, and we can actually distribute tasks and calculation.

However, there's still a problem on the worker side, dealing with the result
update transaction:
    AttributeError: 'Manager' object has no attribute 'select_for_update'

Since we're actually getting to this point, that means that the calculation
(using the nhlib hazard curve calc) concluded "successfully". This is good
progress.