source/simple_fault: fixed a bug in _get_rupture_dimensions() (+)
for the cases when one of dimensions of rupture doesn't exceed
respective dimension of the fault, but other one does, there could be
the situation when after reshaping the previously-fittable dimension
of rupture starts to exceed the one of the fault.

in other words, if for example fault is 10x20 (area is 200) and target
rupture size (according to MSR and aspect ratio) is 12x18 (area is 216),
after reshaping rupture's width (which doesn't fit initially) becomes
10, but length becomes 21.6, which gives more rupture mesh cells along
length than the fault actually has. in the end it results in zero
division error in SimpleFaultSource.iter_ruptures().

now more reliable condition is used for checking if the rupture should
be considered as covering the whole fault area: target rupture area is
compared to total fault surface area.