calculators/hazard/general:
Moved the `execute` function from the classical hazard calculator
into the general base class and made some very small tweaks to the
task function execution in order to make it generic and easily
usable by other calculators.

This also warranted moving a couple of utility functions to the
general module.


Former-commit-id: 1e4a1f1c3579f04c6908a94b4226ef1af7f45a0b