calculators/hazard/general:
Moved the `execute` function from the classical hazard calculator
into the general base class and made some very small tweaks to the
task function execution in order to make it generic and easily
usable by other calculators.

This also warranted moving a couple of utility functions to the
general module.


Former-commit-id: 821002b98741f4583143d2b1b2093b3071d62e27 [formerly 55026de81958864c12c20dd023334565a0cc5fdc] [formerly 55026de81958864c12c20dd023334565a0cc5fdc [formerly 1e4a1f1c3579f04c6908a94b4226ef1af7f45a0b]]
Former-commit-id: 739f2e9dd0ea16e55af86db59067eb9edd966759
Former-commit-id: 7031cc6eafdfe6ad7404ae12b81a2020914a6b64