utils/tasks:
Fixed a very strange bug. Before this work on the event based calculator,
the `hazard_curves` task (in the classical calculator) was the only place
in the codebase to make use of the new @oqtask decorator.

So when I tried to execute the `stochastic_event_sets` tasks (using the
@oqtask decorator), for some reason it would try to execute `hazard_curve`
instead. Very strange.

The solution was to add a functools.wraps decorator to the wrapped function.
Although this fixed the issue, it's not clear from the functools documentation
why this helped. :(