calculators/hazard/classical/core:
Refactored the task arg generation. For the event based calculator,
we'll need to generate the same task args as classical, in addition
to a random seed. These seeds will be generated in sequence using
the calculation `random_seed`.

We will need to apply this random seed (at the beginning of each event
based task) to the global numpy random seed state. Once set, we can
generate SESs and ruptures (the process for which includes a random
element) in a consistent manner (so that the calculation will give the
same results every time).


Former-commit-id: 8189f6146627b566899a8bd320402973cedc1022 [formerly 8189f6146627b566899a8bd320402973cedc1022 [formerly 65edf0c2827980ea65f3f65617d702e6ffe1ed4d]]
Former-commit-id: 1804632edc84a27b8980623c8a60adbaaf94db66
Former-commit-id: abdab1ae816c7f1de588c8324f8c61c489c8db7b