calculators/hazard/classical/core:
Refactored the task arg generation. For the event based calculator,
we'll need to generate the same task args as classical, in addition
to a random seed. These seeds will be generated in sequence using
the calculation `random_seed`.

We will need to apply this random seed (at the beginning of each event
based task) to the global numpy random seed state. Once set, we can
generate SESs and ruptures (the process for which includes a random
element) in a consistent manner (so that the calculation will give the
same results every time).


Former-commit-id: 65edf0c2827980ea65f3f65617d702e6ffe1ed4d