tests/data/event_based_hazard:
Test data for exercising the event-based hazard calculator.

The idea here is that we have 1 of each source typology (area, point,
simple, and complex) with a properly configured magnitude-frequency
distribution such that we will get a good set of ruptures in the given
time span (50 years). This frequency of ruptures is rather higher in
comparison to real-world scenarios, but this is just fine for test cases.


Former-commit-id: 787527802a369eaf44ceae035a73d6ebeb3ce523 [formerly bde1a63e6267e9b61eea63806179ab0200a15f22]
Former-commit-id: 51dfdea4a223fce1d5d38adfedc2e17637ca41be