tests/data/event_based_hazard:
Test data for exercising the event-based hazard calculator.

The idea here is that we have 1 of each source typology (area, point,
simple, and complex) with a properly configured magnitude-frequency
distribution such that we will get a good set of ruptures in the given
time span (50 years). This frequency of ruptures is rather higher in
comparison to real-world scenarios, but this is just fine for test cases.