utils/tasks:
Added a `raise` to the general try/except in the oqtask decorator.

This is helpful, at least in tests: Without this, a test of task code
which encounters an exception with block forever and hang the test.
With this change, the test will fail immediately.

In "production" (with the job supervisor in place), the job supervisor
detects errors by listening to logged error messages. And this is okay.


Former-commit-id: c5ea42c5131e96866df333270d4dcc7791051e56 [formerly c5ea42c5131e96866df333270d4dcc7791051e56 [formerly 4cafd2ff469ac026c34bce03771da523bb0ca287]]
Former-commit-id: 62589b4da195b823a5d7971e67ba85fc3545f5a2
Former-commit-id: 574162595c7d91d00a519f642a3f7c15882cec39