utils/tasks:
Added a `raise` to the general try/except in the oqtask decorator.

This is helpful, at least in tests: Without this, a test of task code
which encounters an exception with block forever and hang the test.
With this change, the test will fail immediately.

In "production" (with the job supervisor in place), the job supervisor
detects errors by listening to logged error messages. And this is okay.


Former-commit-id: 4cafd2ff469ac026c34bce03771da523bb0ca287