openquake/calculators/hazard/classical/core:
Fixed bug where post_process would crash.

If a job is run without any post-processing tasks (such as means
or quantiles), the classical calculator will still attempt to
create and distribute tasks... even when there's no work to be done.

So, I just added a short circuit to skip all of that code if no
post-processing is actually needed.