openquake/calculators/hazard/classical/core:
Fixed bug where post_process would crash.

If a job is run without any post-processing tasks (such as means
or quantiles), the classical calculator will still attempt to
create and distribute tasks... even when there's no work to be done.

So, I just added a short circuit to skip all of that code if no
post-processing is actually needed.


Former-commit-id: b9d1cc9452ec7f4ade076063b5356333e0550981 [formerly b9d1cc9452ec7f4ade076063b5356333e0550981 [formerly 2bb751b3e0ea78ad675620404619387e5394af02]]
Former-commit-id: a04a0b756205c27c46f96987938417acbe65607f
Former-commit-id: 639bea491c89228aac0b563468719f7baf6bd487