utils/tasks, openquake.cfg:
We can now specify the celery work queue in the config file (using the
`celery_queue` parameter). This is useful if there are multiple users
on a single machine, each with its own pool of workers. We can use this
new feature to separate the work simply by running celeryd like so:

$ celeryd -Q foo

and changing `celery_queue` to "foo".


Former-commit-id: 2fecb6a2ab208b439e13037f1b0f0539b1419695 [formerly ad7bddb8398cbe3c94b39c1d9e1465273f73a3be]
Former-commit-id: 8c9ccaff8c92d7bfb8dfce720ebb0c4dcb1bf864