utils/tasks, openquake.cfg:
We can now specify the celery work queue in the config file (using the
`celery_queue` parameter). This is useful if there are multiple users
on a single machine, each with its own pool of workers. We can use this
new feature to separate the work simply by running celeryd like so:

$ celeryd -Q foo

and changing `celery_queue` to "foo".