calcs/hazard/general:
Renamed the `task size/block size` arg for the task arg gen. It
has a more general name now.

The reason why I'm doing this is because, with the addition of the
diaggregation calculation, tasks are no longer just parallelized over
sources. With disagg, parallelization occurs over sources and sites
(sources for hazard curves, sites for disagg histograms).


Former-commit-id: aa106a7350c8ef83416827493a8949bddd264b74