calcs/hazard/general:
Renamed the `task size/block size` arg for the task arg gen. It
has a more general name now.

The reason why I'm doing this is because, with the addition of the
diaggregation calculation, tasks are no longer just parallelized over
sources. With disagg, parallelization occurs over sources and sites
(sources for hazard curves, sites for disagg histograms).


Former-commit-id: 2f93f8801a0d55285841f26e6a6f91b9c93b8766 [formerly 2f93f8801a0d55285841f26e6a6f91b9c93b8766 [formerly aa106a7350c8ef83416827493a8949bddd264b74]]
Former-commit-id: 6f58435d004d2816d61e4bab82f34b6a071f6585
Former-commit-id: 9864c784642603c242d49de207aaeaca04b53911