calcs/hazard/disagg/core:
Added location WKT (point geom) info to disaggregation output
display names.

This helps differentiate results from different locations (when
listing calc outputs on the command line). We want to use the
display as base for an exported filename, so it is actually
quite important that we make this distinction; without the location
info, multiple outputs could have the same name, and that's bad.


Former-commit-id: 51313bb9272ac1fbda907054a29dd67a8f07297b