calcs/hazard/disagg/core:
Added location WKT (point geom) info to disaggregation output
display names.

This helps differentiate results from different locations (when
listing calc outputs on the command line). We want to use the
display as base for an exported filename, so it is actually
quite important that we make this distinction; without the location
info, multiple outputs could have the same name, and that's bad.


Former-commit-id: 2408a9b0e4895942e78b2e529927baf07d785615 [formerly 2408a9b0e4895942e78b2e529927baf07d785615 [formerly 51313bb9272ac1fbda907054a29dd67a8f07297b]]
Former-commit-id: 0b94e08600be7e489ca87cc37d44905b0d1fbd32
Former-commit-id: 474947a4e18bdea06a85e6bc1e12cd8b422a8931