geo/polygon:
Added a fix for cases where dilating certain (extremely convex polygons)
will create holes in the shape and thus result in a multi-polygon.

Dilation usually occurs in the context of generating rupture enclosing
polygons, so the solution here is to simply take the exterior of the resulting
shape, thus discarding the holes. For our nhlib use cases, this is quite
acceptable.