site:
Restore the original numpy-based SiteCollection `expand` algorithm.

While slower than a plain Python implementation with small datasets,
the advantage of numpy really shows through with larger datasets.

I've been meaning to reverse this change for a while, since I found
in profiling run that this change was indeed not faster.