Fixed the --load-curve command
When loading a hazard curve using the `--load-curve` command, the `maximum_distance` parameter is not set, even though it is a mandatory parameter. This fix sets the parameter `maximum_distance` to a default value of 1000 km, which is the same value used when loading a gmf using the `--load-gmf` command.

Former-commit-id: 4167e86e87f9aa83ad974515c6fd49c23760e2af [formerly 498928b608849fd35ae4ce3e119a09582c766a3b] [formerly 498928b608849fd35ae4ce3e119a09582c766a3b [formerly bf4f8cd874021bcb1df4468967cec5ce10a92cd9]]
Former-commit-id: 24866c1c5c345440d53d2eec0f9a4abe92a8b6a7
Former-commit-id: 1ffe6732d8f13efd46543f71c0f743fd828b11f7