Add Attenuation Model of Ghasemi2009
Attenuation model of Ghasemi (2009) is developed for Iran.
A gsim library of this model is added. The changes made are:
1- A gsim library of "ghasemi_2009.py" is added to openquake\hazardlib\gsim
2- A test file of "ghasemi_2009_test.py" is added to openquake\hazardlib\tests\gsim
3- A verification table of "Ghasemi09.xlsx" is added to openquake\hazardlib\tests\gsim\data.
The verification is done using an excel file which calculates the Sa(T1) based on model developed by Ghasemi2009. The results obtained from this excel file is compared with the ones obtained from oq.
4- The "openquake.hazardlib.gsim.rst" file is modified so that it includes "ghasemi_2009" gsim library.
The version and tag of openquake engine used to develope this library is 3.17.1.